Implementation of the Graph logic

+++ app/src/main/java/com/github/vincebrees/bitcoinmarket/data/remote/pojo/RestCoordonate.kt
@@ -5,6 +5,6 @@ package com.github.vincebrees.bitcoinmarket.data.remote.pojo
  */
 
 data class RestCoordonate(
-    var x : Double,
+    var x : Long,
     var y : Double
 )
