+++ app/src/main/java/com/github/vincebrees/bitcoinmarket/domain/entity/Coordonate.kt
@@ -5,6 +5,6 @@ package com.github.vincebrees.bitcoinmarket.domain.entity
  */
 
 data class Coordonate(
-    var x : Double,
+    var x : Long,
     var y : Double
 )
